feat: introduce PredictionResolver service for automated prediction resolution
- Move prediction resolution logic from AutoClient to a dedicated PredictionResolver service
- Add new ServiceType.PREDICTION_RESOLVER to core types
- Implement interval-based prediction resolution with weather data retrieval and outcome determination
- Improve logging and error handling for prediction resolution process
